requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
pinecone
pinecone_text
sentence-transformers
duckduckgo-search
langchain-experimental
speech-recognition
pyttsx3
pyaudio
whisper
ffmpeg-python
elevenlabs==0.3.0b0
rpaudio
PIL
io
pyt2s
time